Move styling examples to styling/ folder
