+++ examples/styling/jss/src/components/Profile/index.js
