+++ package.json
@@ -39,9 +39,9 @@
     "example:dest": "npm run server -- --context examples/dest/ --config examples/dest/webpack.dev.babel.js",
     "example:dest:iron": "iron-node node_modules/webpack-dev-server/bin/webpack-dev-server.js --context examples/dest/ --config examples/dest/webpack.dev.babel.js",
     "example:profile": "npm run server -- --context examples/profile/ --config examples/profile/webpack.dev.babel.js",
-    "example:radium": "npm run server -- --context examples/radium/ --config examples/radium/webpack.dev.babel.js",
-    "example:jss": "npm run server -- --context examples/jss/ --config examples/jss/webpack.dev.babel.js",
-    "example:aphrodite": "npm run server -- --context examples/aphrodite/ --config examples/aphrodite/webpack.dev.babel.js",
+    "example:radium": "npm run server -- --context examples/styling/radium/ --config examples/styling/radium/webpack.dev.babel.js",
+    "example:jss": "npm run server -- --context examples/styling/jss/ --config examples/styling/jss/webpack.dev.babel.js",
+    "example:aphrodite": "npm run server -- --context examples/styling/aphrodite/ --config examples/styling/aphrodite/webpack.dev.babel.js",
     "client:dev": "webpack-dev-server --hot --config client/webpack.dev.babel.js",
     "client:build": "webpack --config client/webpack.prod.babel.js --progress -p",
     "build": "npm run client:build",
